feat: [UserManagement/Frontend] extracted table rows and header of UserManagementPage into UserTableRow
