+++ frontend/src/pages/UserManagement/UserManagementPage.tsx
@@ -2,8 +2,8 @@ import "./UserManagementPage.css"
 import {useEffect, useState} from "react";
 import {BackendAPI} from "../../global_functions/BackendAPI.tsx";
 import {StoredUserInfo} from "../../models/UserManagementTypes.tsx";
-import BreadCrumbs from "../../components/BreadCrumbs.tsx";
 import styled from "styled-components";
+import UserTableRow from "./components/UserTableRow.tsx";
 
 const TableCard = styled.div`
   border: 1px solid var(--border-color, #707070);
@@ -23,52 +23,19 @@ export default function UserManagementPage() {
         );
     }, []);
 
-    function getReplacementIfNeeded(str: string, replacement: string) {
-        if (!str || str.trim()==="")
-            return replacement;
-        return str;
-    }
-
     return (
         <>
             <h3>User Management</h3>
-            <BreadCrumbs/>
             <TableCard>
                 <table className={"UserTable"}>
                     <thead>
-                        <tr>
-                            <th>User           </th>
-                            <th>Role           </th>
-                            <th>Deny Access    </th>
-                            <th>Login Name     </th>
-                            <th>Location       </th>
-                            <th>URL            </th>
-                            <th>ID             </th>
-                            <th>Registration ID</th>
-                            <th>Original ID    </th>
-                        </tr>
+                        <UserTableRow/>
                     </thead>
                     <tbody>
                     {
-                        users.map( user => {
-                            return (
-                                <tr key={user.id}>
-                                    <td className={"Name"}>
-                                        {user.avatar_url && <img className={"AvatarImage"} alt={"Avatar of user with ID "+user.id} src={user.avatar_url}/>}
-                                        {"   "}
-                                        {getReplacementIfNeeded(user.name, "------")}
-                                    </td>
-                                    <td>{user.role          }</td>
-                                    <td className={"DenialReason"}>{user.denialReason  }</td>
-                                    <td>{user.login         }</td>
-                                    <td>{user.location      }</td>
-                                    <td><a href={user.url}>{user.url}</a> </td>
-                                    <td>{user.id            }</td>
-                                    <td>{user.registrationId}</td>
-                                    <td>{user.originalId    }</td>
-                                </tr>
-                            )
-                        })
+                        users.map( user =>
+                            <UserTableRow key={user.id} user={user}/>
+                        )
                     }
                     </tbody>
                 </table>
